% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly2DailyQ.R
\name{hourly2DailyQ}
\alias{hourly2DailyQ}
\title{Aggregate NetCDF Hourly Streamflow to Daily}
\usage{
hourly2DailyQ(directory, comids)
}
\arguments{
\item{directory}{Directory where hourly NetCDF files are kept. Must be named for the year (yyyy) of the NetCDF files contained inside.}

\item{comids}{A vector of COMID values (identifiers of NHDPlus stream reaches).}
}
\value{
A time series of daily streamflow for each of the user-specified reaches.
}
\description{
Aggregate NetCDF Hourly Streamflow to Daily
}
\examples{
dailyQ <- hourly2DailyQ("~/1994",c("18578829", "18578755"))
}
