% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowFlowEval.R
\name{lowFlowEval}
\alias{lowFlowEval}
\title{Compare number of low flow events between NWM and USGS records}
\usage{
lowFlowEval(gageID, comID, flowfile, threshold, startDate, endDate)
}
\arguments{
\item{gageID}{USGS gage ID, string}

\item{comID}{a reach ID (corresponds to the feature IDs in the NHD Plus dataset), string}

\item{flowfile}{.csv file of time series data, with columns of Date and comIDs}

\item{threshold}{a measure of low flow ("Perc5", "Perc25", "x7Q2", or "x7Q10"), string}

\item{startDate}{beginning date for data retrieval, "yyyy-mm-dd"}

\item{endDate}{end date for data retrieval, "yyyy-mm-dd"}
}
\value{
A time series of daily streamflow for each of the user-specified reaches.
}
\description{
Compare number of low flow events between NWM and USGS records
}
\examples{


}
