% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFlowStats.R
\name{NWMFlowStats}
\alias{NWMFlowStats}
\title{NWMFlowStats}
\usage{
NWMFlowStats(comID, flowfile)
}
\arguments{
\item{comID}{a reach ID (corresponds to the feature IDs in the NHD Plus dataset)}

\item{flowfile}{.csv file of time series data, with columns of Date and comIDs}
}
\value{
dataframe of monthly flow statistics
}
\description{
NWMFlowStats calculates stats for a given reach, using a user-provided .csv file of NWM data.
}
