% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroNWMUSGSEval.R
\name{retroNWMUSGSEval}
\alias{retroNWMUSGSEval}
\title{retroNWMUSGSEval}
\usage{
retroNWMUSGSEval(gageID, comID, flowfile, plot.on = TRUE)
}
\arguments{
\item{gageID}{USGS gage ID, character}

\item{comID}{beginning date for data retrieval, "yyyy-mm-dd"}

\item{flowfile}{end date for data retrieval, "yyyy-mm-dd"}

\item{plot.on}{plots time series of input data. Default is TRUE}
}
\value{
list of goodness of fit statistics
}
\description{
Calculates a suite of goodness of fit/model performance statistics. Calculates overall GOF,
seasonal, and by year.
}
