% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFlowStats.R
\name{USGSFlowStats}
\alias{USGSFlowStats}
\title{USGSFlowStats}
\usage{
USGSFlowStats(gageID, startDate, endDate)
}
\arguments{
\item{gageID}{USGS gage ID, character}

\item{startDate}{beginning date for data retrieval, "yyyy-mm-dd"}

\item{endDate}{end date for data retrieval, "yyyy-mm-dd"}
}
\value{
dataframe of monthly flow statistics
}
\description{
USGSFlowStats calculates summary stats for streamflow at a given stream gage using the USGS dataRetrieval package.
}
