% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly2DailyQ.R
\name{hourly2DailyQ}
\alias{hourly2DailyQ}
\title{Aggregate NetCDF Hourly Streamflow to Daily}
\usage{
hourly2DailyQ(directory, comids, suffix = "", write.output = TRUE)
}
\arguments{
\item{directory}{Directory where hourly NetCDF files are kept. Must be named for the year (yyyy) of the NetCDF files contained inside.}

\item{comids}{A vector of COMID values (identifiers of NHDPlus stream reaches).}

\item{suffix}{Optional suffix to append to the output file name. Default is no suffix.}

\item{write.output}{Boolean for whether or not to write the results to a .csv file. Default is TRUE.}
}
\value{
A time series of daily streamflow for each of the user-specified reaches.
}
\description{
Aggregate NetCDF Hourly Streamflow to Daily
}
\examples{
dailyQ <- aggHourlyStreamflow("~/1994","8020924","CaseStudy",TRUE)
}
